Added all calibration fields that sends strings to arduino
The user has all calibration fields now that will create a string in the format Sam specified in his picture. It will also send to arduino using the line, "btSocket.getOutputStream().write(str.toString().getBytes());"

Please check ledControl.java's txtField# and tog listeners to see implementation on how string is setup and sent to arduino. I do not know if this is the format that the arduino code wants and should be changed to be accepted.